refactor: troca tipo de dado da propriedade Price

+++ internal/entity/product.go
@@ -18,11 +18,11 @@ var (
 type Product struct {
 	ID        entity.ID `json:"id"`
 	Name      string    `json:"name"`
-	Price     float32   `json:"price"`
+	Price     float64   `json:"price"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
-func NewProduct(name string, price float32) (*Product, error) {
+func NewProduct(name string, price float64) (*Product, error) {
 	p := &Product{
 		ID:        entity.NewID(),
 		Name:      name,
